post

+++ app/Http/Controllers/TasksController.php
@@ -15,11 +15,24 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            
+        }
+    
+        //return view('tasks.index', ['tasks' => $tasks]);
         
-        return view('tasks.index',[
-            'tasks' => $tasks,
-        ]);
+        return view('tasks.index', $data);
+        
+        
+
     }
 
     /**
@@ -29,6 +42,7 @@ class TasksController extends Controller
      */
     public function create()
     {
+        
         $task = new Task;
         
         return view('tasks.create', [
@@ -49,10 +63,16 @@ class TasksController extends Controller
             'status' => 'required|max:10',
         ]);
         
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        //$task = new Task;
+        //$task->status = $request->status;
+        //$task->content = $request->content;
+        //$task->save();
+        
+        
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
         
         return redirect('/');
     }
@@ -67,9 +87,15 @@ class TasksController extends Controller
     {
         $task = Task::find($id);
         
-        return view('tasks.show',[
-            'task' => $task,
-        ]);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.show',[
+                'task' => $task,
+            ]);
+        }
+        
+        else {
+            return redirect('/');
+        }
     }
 
     /**
@@ -82,9 +108,15 @@ class TasksController extends Controller
     {
         $task = Task::find($id);
         
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+        }
+        
+        else {
+             return redirect('/');
+        }
     }
 
     /**
@@ -118,8 +150,13 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
         
-        return redirect('/');
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+        
+        else {
+             return redirect('/');
+        }
     }
 }
